more simplifcation
