+++ packages/rag/src/document/extractors/index.ts
@@ -2,4 +2,4 @@ export { TitleExtractor } from './title';
 export { SummaryExtractor } from './summary';
 export { QuestionsAnsweredExtractor } from './questions';
 export { KeywordExtractor } from './keywords';
-export * from './types';
+export type { KeywordExtractArgs, QuestionAnswerExtractArgs, SummaryExtractArgs, TitleExtractorsArgs } from './types';
