+++ packages/rag/src/document/prompts/index.ts
@@ -1,5 +1,4 @@
 export { BasePromptTemplate, PromptTemplate } from './base';
-export type { BasePromptTemplateOptions, PromptTemplateOptions, StringTemplate } from './base';
 export {
   defaultKeywordExtractPrompt,
   defaultQuestionExtractPrompt,
